# BinaryWriter perf and memory improvements