… src/libraries/System.IO/tests/BinaryWriter/BinaryWriter.EncodingTests.cs