This addresses some low-hanging fruit in the `BinaryWriter` class, reducing overall memory footprint and wall clock time for common operations. It also removed use of the _unsafe_ keyword where possible.

I spoke offline with @adamsitnik about the consequences behind changing a bunch of `Steam.Write(byte[], int, int)` call sites to read `Stream.Write(ROS<byte>)` instead. Technically this could result in worse performance if the wrapped stream doesn't override the `ROS<byte>`-based overloads, since the default implementations of those overloads will rent from the array pool, copy, and forward to the array-based overloads. But honestly, it's 2021, most of the built-in stream types override these methods correctly, and we're already discussing ways to flag with warnings user-defined types which don't override these methods. I don't think we should handicap the common case of using fully-compliant built-in stream types just on the off-chance somebody might have used a custom type.

Perf results:

|      Method |        Job | Toolchain |      Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------- |---------- |----------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| DefaultCtor | Job-LUELTV |    master | 28.972 ns | 0.5700 ns | 0.5332 ns |  1.00 | 0.0181 |     - |     - |     152 B |
| DefaultCtor | Job-GIFVTY |        pr | 13.240 ns | 0.3233 ns | 0.3594 ns |  0.46 | 0.0048 |     - |     - |      40 B |
|             |            |           |           |           |           |       |        |       |       |           |
| WriteUInt32 | Job-LUELTV |    master |  2.665 ns | 0.0159 ns | 0.0133 ns |  1.00 |      - |     - |     - |         - |
| WriteUInt32 | Job-GIFVTY |        pr |  2.090 ns | 0.0352 ns | 0.0294 ns |  0.78 |      - |     - |     - |         - |
|             |            |           |           |           |           |       |        |       |       |           |
| WriteUInt64 | Job-LUELTV |    master |  1.876 ns | 0.0130 ns | 0.0109 ns |  1.00 |      - |     - |     - |         - |
| WriteUInt64 | Job-GIFVTY |        pr |  2.101 ns | 0.0137 ns | 0.0121 ns |  1.12 |      - |     - |     - |         - |

|         Method |        Job | Toolchain | StringLengthInChars |          Mean |         Error |        StdDev |        Median | Ratio | RatioSD |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|--------------- |----------- |---------- |-------------------- |--------------:|--------------:|--------------:|--------------:|------:|--------:|---------:|---------:|---------:|----------:|
| **WriteCharArray** | **Job-XZHCJM** |    **master** |                   **4** |      **24.38 ns** |      **0.208 ns** |      **0.194 ns** |      **24.35 ns** |  **1.00** |    **0.00** |   **0.0038** |        **-** |        **-** |      **32 B** |
| WriteCharArray | Job-EAPOXA |        pr |                   4 |      35.05 ns |      0.113 ns |      0.100 ns |      35.02 ns |  1.44 |    0.01 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |                   4 |      20.39 ns |      0.066 ns |      0.058 ns |      20.39 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |                   4 |      14.80 ns |      0.034 ns |      0.026 ns |      14.81 ns |  0.73 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |                  **16** |      **27.92 ns** |      **0.176 ns** |      **0.165 ns** |      **27.93 ns** |  **1.00** |    **0.00** |   **0.0048** |        **-** |        **-** |      **40 B** |
| WriteCharArray | Job-EAPOXA |        pr |                  16 |      35.11 ns |      0.115 ns |      0.108 ns |      35.14 ns |  1.26 |    0.01 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |                  16 |      24.15 ns |      0.088 ns |      0.078 ns |      24.13 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |                  16 |      16.96 ns |      0.111 ns |      0.098 ns |      16.93 ns |  0.70 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |                 **512** |      **75.80 ns** |      **0.923 ns** |      **0.864 ns** |      **75.51 ns** |  **1.00** |    **0.00** |   **0.0640** |        **-** |        **-** |     **536 B** |
| WriteCharArray | Job-EAPOXA |        pr |                 512 |      52.93 ns |      0.242 ns |      0.227 ns |      52.92 ns |  0.70 |    0.01 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |                 512 |     137.70 ns |      0.903 ns |      0.845 ns |     137.74 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |                 512 |      52.54 ns |      0.537 ns |      0.449 ns |      52.53 ns |  0.38 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |               **10000** |   **1,044.99 ns** |     **26.915 ns** |     **78.936 ns** |   **1,023.76 ns** |  **1.00** |    **0.00** |   **1.1959** |        **-** |        **-** |   **10024 B** |
| WriteCharArray | Job-EAPOXA |        pr |               10000 |     358.84 ns |      1.177 ns |      1.044 ns |     358.70 ns |  0.35 |    0.02 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |               10000 |   2,748.17 ns |     24.340 ns |     22.767 ns |   2,748.62 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |               10000 |     352.63 ns |      1.601 ns |      1.498 ns |     353.18 ns |  0.13 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |              **100000** |  **39,499.04 ns** |    **784.557 ns** |  **1,894.794 ns** |  **39,615.28 ns** |  **1.00** |    **0.00** |  **31.1890** |  **31.1890** |  **31.1890** |  **100025 B** |
| WriteCharArray | Job-EAPOXA |        pr |              100000 |   3,193.04 ns |     18.645 ns |     17.441 ns |   3,192.28 ns |  0.08 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |              100000 |  27,025.32 ns |    139.589 ns |    123.742 ns |  26,993.91 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |              100000 |   3,483.60 ns |     20.989 ns |     19.633 ns |   3,485.38 ns |  0.13 |    0.00 |        - |        - |        - |         - |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |              **500000** |  **90,273.26 ns** |  **1,036.222 ns** |    **918.584 ns** |  **90,217.61 ns** |  **1.00** |    **0.00** | **113.7695** | **113.7695** | **113.7695** |  **500021 B** |
| WriteCharArray | Job-EAPOXA |        pr |              500000 |  17,491.07 ns |     69.793 ns |     65.285 ns |  17,526.21 ns |  0.19 |    0.00 |        - |        - |        - |      48 B |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |              500000 | 137,453.00 ns |    649.887 ns |    607.905 ns | 137,337.48 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |              500000 |  30,467.77 ns |    182.114 ns |    170.350 ns |  30,530.05 ns |  0.22 |    0.00 |        - |        - |        - |      48 B |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
| **WriteCharArray** | **Job-XZHCJM** |    **master** |             **2000000** | **766,902.63 ns** | **15,059.390 ns** | **23,445.656 ns** | **769,048.83 ns** |  **1.00** |    **0.00** |  **95.7031** |  **95.7031** |  **95.7031** | **2000032 B** |
| WriteCharArray | Job-EAPOXA |        pr |             2000000 |  71,244.09 ns |    236.232 ns |    209.414 ns |  71,242.82 ns |  0.09 |    0.00 |        - |        - |        - |     160 B |
|                |            |           |                     |               |               |               |               |       |         |          |          |          |           |
|    WriteString | Job-XZHCJM |    master |             2000000 | 645,989.09 ns |  4,823.895 ns |  4,512.275 ns | 647,063.96 ns |  1.00 |    0.00 |        - |        - |        - |         - |
|    WriteString | Job-EAPOXA |        pr |             2000000 | 126,507.40 ns |    719.938 ns |    673.430 ns | 126,473.10 ns |  0.20 |    0.00 |        - |        - |        - |     160 B |

The _WriteChars_ tests for small values requires further discussion. The original implementation allocates a new array on each invocation, while the new implementation uses the array pool. The indirection through the array pool adds a few nanoseconds fixed overhead, which causes the ratio difference between the old and new code to be exaggerated. I believe the new code is more appropriate for the common case since it reduces the overall memory footprint of the application, even with this overhead. There is also some overhead due to the delegate invocation in the workhorse routine. When the delegate is first created, it points to a stub routine rather than directly to the target method, adding a few extra jumps. This is a long-standing behavioral nit in delegates and if it's solved all-up in the runtime then we'll just get the benefits here for free.

Benchmark code below.

<details>
<p>

```cs
using BenchmarkDotNet.Attributes;
using System;
using System.IO;
using System.Threading;
using System.Threading.Tasks;

namespace ConsoleAppBenchmark
{
    [MemoryDiagnoser]
    public class BinaryWriterRunner
    {
        private BinaryWriter _bw;

        [GlobalSetup]
        public void Setup()
        {
            _bw = new BinaryWriter(new NullWriteStream());
        }

        [Benchmark]
        public BinaryWriter DefaultCtor() => new BinaryWriter(Stream.Null);

        [Benchmark]
        public void WriteUInt32()
        {
            _bw.Write((uint)0xdeadbeef);
        }

        [Benchmark]
        public void WriteUInt64()
        {
            _bw.Write((ulong)0xdeadbeef_aabbccdd);
        }
    }

    [MemoryDiagnoser]
    public class BinaryWriterRunner_Extended
    {
        private string _input;
        private char[] _inputAsChars;
        private readonly BinaryWriter _bw;

        [Params(4, 16, 512, 10_000, 100_000, 500_000, 2_000_000)]
        public int StringLengthInChars;

        public BinaryWriterRunner_Extended()
        {
            _bw = new BinaryWriter(new NullWriteStream());
        }

        [GlobalSetup]
        public void Setup()
        {
            _input = new string('x', StringLengthInChars);
            _inputAsChars = _input.ToCharArray();
        }

        [Benchmark]
        public void WriteCharArray()
        {
            _bw.Write(_inputAsChars);
        }

        [Benchmark]
        public void WriteString()
        {
            _bw.Write(_input);
        }
    }

    internal class NullWriteStream : Stream
    {
        public override bool CanRead => false;

        public override bool CanSeek => false;

        public override bool CanWrite => true;

        public override long Length => throw new NotSupportedException();

        public override long Position { get => throw new NotSupportedException(); set => throw new NotSupportedException(); }

        public override void Flush() { }

        public override int Read(byte[] buffer, int offset, int count)
        {
            throw new NotSupportedException();
        }

        public override long Seek(long offset, SeekOrigin origin)
        {
            throw new NotSupportedException();
        }

        public override void SetLength(long value)
        {
            throw new NotSupportedException();
        }

        public override void Write(byte[] buffer, int offset, int count) { }

        public override void Write(ReadOnlySpan<byte> buffer) { }

        public override Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken)
        {
            return Task.CompletedTask;
        }

        public override void WriteByte(byte value) { }

        public override ValueTask WriteAsync(ReadOnlyMemory<byte> buffer, CancellationToken cancellationToken = default)
        {
            return ValueTask.CompletedTask;
        }
    }
}
```

</p>
</details>