@GrabYourPitchforks 2021-01-27
@adamsitnik I sent a PR at https://github.com/dotnet/performance/pull/1639 with these tests, plus added the tests you suggested + fleshed out a few others.