@GrabYourPitchforks 2021-01-27
@adamsitnik @jkotas I didn't see much of a difference between using a 32k vs. a 64k max rental size. Perf results below.

|         Method | Toolchain | StringLengthInChars |         Mean |        Error |     StdDev | Ratio | RatioSD |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|--------------- |---------- |-------------------- |-------------:|-------------:|-----------:|------:|--------:|-------:|------:|------:|----------:|
| **WriteCharArray** |     **pr32k** |                   **4** |     **33.00 ns** |     **1.138 ns** |   **0.062 ns** |  **0.99** |    **0.01** |      **-** |     **-** |     **-** |         **-** |
| WriteCharArray |     pr64k |                   4 |     33.27 ns |     2.128 ns |   0.117 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |                   4 |     12.04 ns |     0.612 ns |   0.034 ns |  0.99 |    0.00 |      - |     - |     - |         - |
|    WriteString |     pr64k |                   4 |     12.12 ns |     1.073 ns |   0.059 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |                  **16** |     **34.69 ns** |     **0.626 ns** |   **0.034 ns** |  **1.00** |    **0.01** |      **-** |     **-** |     **-** |         **-** |
| WriteCharArray |     pr64k |                  16 |     34.85 ns |     7.338 ns |   0.402 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |                  16 |     13.27 ns |     0.396 ns |   0.022 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|    WriteString |     pr64k |                  16 |     13.23 ns |     0.745 ns |   0.041 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |                 **512** |     **59.53 ns** |     **5.954 ns** |   **0.326 ns** |  **1.00** |    **0.04** |      **-** |     **-** |     **-** |         **-** |
| WriteCharArray |     pr64k |                 512 |     59.61 ns |    37.954 ns |   2.080 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |                 512 |     51.82 ns |     4.056 ns |   0.222 ns |  0.99 |    0.01 |      - |     - |     - |         - |
|    WriteString |     pr64k |                 512 |     52.21 ns |     2.390 ns |   0.131 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |                **8192** |    **302.99 ns** |     **9.689 ns** |   **0.531 ns** |  **1.02** |    **0.00** |      **-** |     **-** |     **-** |         **-** |
| WriteCharArray |     pr64k |                8192 |    295.62 ns |     8.826 ns |   0.484 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |                8192 |    296.68 ns |    47.442 ns |   2.600 ns |  0.93 |    0.11 |      - |     - |     - |         - |
|    WriteString |     pr64k |                8192 |    320.96 ns |   701.383 ns |  38.445 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |               **16384** |    **632.55 ns** |     **9.539 ns** |   **0.523 ns** |  **1.03** |    **0.01** | **0.0057** |     **-** |     **-** |      **48 B** |
| WriteCharArray |     pr64k |               16384 |    614.72 ns |   156.201 ns |   8.562 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |               16384 |  1,054.91 ns |    93.902 ns |   5.147 ns |  1.67 |    0.01 | 0.0057 |     - |     - |      48 B |
|    WriteString |     pr64k |               16384 |    630.59 ns |    21.631 ns |   1.186 ns |  1.00 |    0.00 |      - |     - |     - |         - |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |              **131072** |  **4,646.23 ns** |   **182.919 ns** |  **10.026 ns** |  **1.05** |    **0.00** | **0.0153** |     **-** |     **-** |     **160 B** |
| WriteCharArray |     pr64k |              131072 |  4,427.34 ns |   126.210 ns |   6.918 ns |  1.00 |    0.00 | 0.0153 |     - |     - |     160 B |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |              131072 |  8,025.57 ns |   377.238 ns |  20.678 ns |  1.00 |    0.00 | 0.0153 |     - |     - |     160 B |
|    WriteString |     pr64k |              131072 |  8,025.43 ns |   364.648 ns |  19.988 ns |  1.00 |    0.00 | 0.0153 |     - |     - |     160 B |
|                |           |                     |              |              |            |       |         |        |       |       |           |
| **WriteCharArray** |     **pr32k** |             **1048576** | **38,643.52 ns** | **9,283.063 ns** | **508.836 ns** |  **1.03** |    **0.02** |      **-** |     **-** |     **-** |     **160 B** |
| WriteCharArray |     pr64k |             1048576 | 37,528.87 ns | 6,376.444 ns | 349.514 ns |  1.00 |    0.00 |      - |     - |     - |     160 B |
|                |           |                     |              |              |            |       |         |        |       |       |           |
|    WriteString |     pr32k |             1048576 | 68,994.16 ns | 4,548.266 ns | 249.306 ns |  1.03 |    0.01 |      - |     - |     - |     160 B |
|    WriteString |     pr64k |             1048576 | 67,256.93 ns | 3,381.513 ns | 185.352 ns |  1.00 |    0.00 |      - |     - |     - |     161 B |

The strlen = 8192, 64kb buffer test has a large stddev, so I'm not worrying too much about it. The strlen = 16384 test is very different between 32kb and 64kb because on 64kb the data fits into a single buffer, while on 32kb it does not so we need to go down the slow "two-pass" path.

I think this means we can stick with 64k.